refactor(mime): sink format_id_default_mime into uc-platform, drop UTI branches from core
The platform → RFC-MIME translation table is a desktop-specific
piece of knowledge (macOS UTIs, NSPasteboard legacy names, Windows
short tags, project-internal Linux tags). It does not belong in
`uc-core`, which under its own AGENTS.md §2.2 bans platform
identifiers from the domain layer.

Changes:

- New `uc_platform::clipboard::format_id_mime::format_id_default_mime`.
  Function body is identical to the previous core implementation; only
  the location moves. Tests come with it.
- `uc-platform/src/clipboard/{common,platform/macos,platform/windows}.rs`
  now import the function from the local platform module instead of
  `uc_core::clipboard`. Linux Wayland/X11 capture paths never called it
  (they already produce IANA MIME directly) — no changes there.
- `format_id_default_mime` removed from `uc-core/src/clipboard/mime.rs`
  and from the `uc-core::clipboard` public re-exports.
- `MimeClass::from_essence` no longer recognizes `public.utf8-plain-text`
  / `public.text` as plain text. The doc comment now states explicitly
  that the engine only accepts RFC media types — capture adapters must
  translate at the platform boundary (a contract Commit B locked down
  for `ObservedClipboardRepresentation::mime`).

Future iOS and Android hosts will add their own translation tables
in their own `uc-platform-*` crates; the engine has no `public.*` /
UTI knowledge to migrate.

Test deltas:
- uc-core: 117 → 116 (moved `format_id_default_mime_handles_known_ids`)
- uc-platform: 38 → 39 (received the same test)
- The `classify_text_plain_variants` case in core dropped the
  `"public.utf8-plain-text"` row, mirroring the from_essence cleanup.
- uc-application 568 tests untouched and green.

This is Commit C of the engine/mobile portability cleanup. The final
step (D) will add UTI→RFC normalization at the wire deserialization
boundary so the application-layer defensive checks on `mime` can be
deleted without breaking inbound sync from older peers.
