+++ src-tauri/crates/uc-platform/src/clipboard/platform/macos.rs
@@ -14,10 +14,11 @@ use objc2_foundation::{NSArray, NSData};
 use std::sync::{Arc, Mutex};
 use tracing::{debug, debug_span, info, warn};
 use uc_core::clipboard::{
-    format_id_default_mime, ImageKind, MimeClass, MimeType, ObservedClipboardRepresentation,
-    SystemClipboardSnapshot,
+    ImageKind, MimeClass, MimeType, ObservedClipboardRepresentation, SystemClipboardSnapshot,
 };
 use uc_core::ports::SystemClipboardPort;
+
+use crate::clipboard::format_id_mime::format_id_default_mime;
 
 /// macOS clipboard implementation using clipboard-rs
 pub struct MacOSClipboard {
